% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func1.R
\name{func1}
\alias{func1}
\title{Print the input argument}
\usage{
func1(x)
}
\arguments{
\item{x}{:A numeric or character value}
}
\value{
print statement with the input value
}
\description{
Print the input argument
}
\examples{
func1(2)
The variable is 2
}
